Add a bunch of places, change some things

+++ districtr_process/columns.py
@@ -1,6 +1,6 @@
 import pandas
 from marshmallow import Schema, fields, post_load
-from pandas.api.types import is_numeric_dtype
+from pandas.api.types import is_integer_dtype, is_numeric_dtype
 
 from .exceptions import MissingColumnsError
 
@@ -34,7 +34,7 @@ class Column:
             raise MissingColumnsError("Missing column {}".format(self))
 
         if self.require_numeric is True:
-            df[self.key] = pandas.to_numeric(df[self.key])
+            df[self.key] = pandas.to_numeric(df[self.key]).astype(int)
 
         column = df[self.key]
